case문 추가

+++ script/ups_read.sh
@@ -1,23 +1,50 @@
 #!/bin/bash
 
 #/*---------------------------------------------------------------------------*/
-#/* find CH55xduino ttyACM node (1209:c550) */
+#/* Define CH55xduino ttyACM VID/PID (1209:c550) */
 #/*---------------------------------------------------------------------------*/
 VID_CH55xduino="1209"
 PID_CH55xduino="C550"
 
 #/*---------------------------------------------------------------------------*/
+#/* Define tmp file */
+#/*---------------------------------------------------------------------------*/
 UPS_TTY_NODE=""
 UPS_TTY_DATA="/tmp/ttyUPS.dat"
 
 #/*---------------------------------------------------------------------------*/
-UPS_STATUS_VOLT="0"
+#/* UPS Command List */
+#/*---------------------------------------------------------------------------*/
+# Send command to read battery volt to UPS.
+UPS_CMD_BATTERY_VOLT="@V0#"
+# Send command to read battery level to UPS.
+# 0 : BATTERY LEVEL 0 (3400 mV)
+# 1 : BATTERY LEVEL 1 (3550 mV)
+# 2 : BATTERY LEVEL 2 (3650 mV)
+# 3 : BATTERY LEVEL 3 (3750 mV)
+# 4 : BATTERY LEVEL 4 (3900 mV)
+UPS_CMD_BATTERY_LEVEL="@L0#"
+# Send command to read charger status to UPS.
+UPS_CMD_BATTERY_STATUS="@C0#"
+# Send command to ups off to UPS.
+UPS_CMD_POWEROFF="@P0#"
+# Send command to power on level to UPS.
+# 0 ~ 4 : BATTERY LEVEL
+# * : Detect charging status.(default)
+UPS_CMD_POWERON="@O0#"
+
+#/*---------------------------------------------------------------------------*/
+UPS_CMD_STR=""
+
+#/*---------------------------------------------------------------------------*/
+UPS_BATTERY_VOLT="0"
 UPS_STATUS_CHRG="0"
 UPS_STATUS_FULL="0"
 
 #/*---------------------------------------------------------------------------*/
 #/* UPS Battery Level define value */
 #/*---------------------------------------------------------------------------*/
+BAT_LEVEL_FULL="4300"
 BAT_LEVEL_LV4="3900"
 BAT_LEVEL_LV3="3750"
 BAT_LEVEL_LV2="3650"
@@ -25,13 +52,21 @@ BAT_LEVEL_LV1="3550"
 
 #/* System force power level */
 #POWER_OFF_VOLT=${BAT_LEVEL_LV3}
-POWER_OFF_VOLT=${BAT_LEVEL_LV4}
+
+#/* Power off when battery discharge condition detected. */
+POWER_OFF_VOLT=${BAT_LEVEL_FULL}
+
+UPS_SEND_CMD=""
+UPS_SEND_STR=""
+UPS_ON_LEVEL=""
+UPS_WATCHDOG_TIME=""
 
 #/*---------------------------------------------------------------------------*/
 #/*---------------------------------------------------------------------------*/
 function kill_dead_process {
+	PID=""
 	PID=`ps -eaf | grep ${UPS_TTY_NODE} | grep -v grep | awk '{print $2}'`
-	if [[ "" !=  "$PID" ]]; then
+	if [ -n "$PID" ]; then
 		echo "------------------------------------------------------------"
 		echo "Killing $PID"
 		kill -9 $PID
@@ -49,42 +84,91 @@ function find_tty_node {
 }
 
 #/*---------------------------------------------------------------------------*/
-function read_ups_volt {
-	# ttyACM response data background setup.
+function ups_cmd_send {
+	# ttyACM response data wait settings.
 	cat ${UPS_TTY_NODE} > ${UPS_TTY_DATA} &
+	sleep 1
 
-	# Get PID(cat command) for kill background process
+	# Get PID (cat command) to kill background process
+	PID=""
 	PID=$!
 
-	# UPS read volt command send
-	sleep 0.5
+	#/* Send command string to UPS */
+	echo -ne ${UPS_CMD_STR} > ${UPS_TTY_NODE}
+	sleep 1
+
+	#/* Update data */
+	case ${UPS_CMD_STR} in
+		${UPS_CMD_BATTERY_VOLT})
+			# Update battery volt data.
+			UPS_BATTERY_VOLT=`cut -c 3-6 < ${UPS_TTY_DATA}`
+			;;
+		${UPS_CMD_BATTERY_LEVEL})
+			# Update charger status data.
+			UPS_BATTERY_LV4=`cut -c 3 < ${UPS_TTY_DATA}`
+			UPS_BATTERY_LV3=`cut -c 4 < ${UPS_TTY_DATA}`
+			UPS_BATTERY_LV2=`cut -c 5 < ${UPS_TTY_DATA}`
+			UPS_BATTERY_LV1=`cut -c 6 < ${UPS_TTY_DATA}`
+			;;
+		${UPS_CMD_CHARGER_STATUS})
+			# Update charger status data.
+			UPS_STATUS_CHRG=`cut -c 6 < ${UPS_TTY_DATA}`
+			UPS_STATUS_FULL=`cut -c 4 < ${UPS_TTY_DATA}`
+			;;
+		* )
+			;;
+	esac
+
+	# Kill background process(cat cmd)
+	if [ -n "$PID" ]; then
+		kill $PID
+	fi
+}
+
+#/*---------------------------------------------------------------------------*/
+#/*---------------------------------------------------------------------------*/
+function read_ups_volt {
+	# ttyACM response data wait settings.
+	cat ${UPS_TTY_NODE} > ${UPS_TTY_DATA} &
+
+	# Get PID (cat command) to kill background process
+	PID=""
+	PID=$!
+
+	# Send command to read battery volt to UPS.
+	sleep 1
 	echo -ne "@V0#" > ${UPS_TTY_NODE}
-	sleep 0.5
+	sleep 1
 
-	# Kill Background process(cat cmd)
-	kill $PID
+	# Kill background process(cat cmd)
+	if [ -n "$PID" ]; then
+		kill $PID
+	fi
 
-	# UPS Volt data update
-	UPS_STATUS_VOLT=`cut -c 3-6 < ${UPS_TTY_DATA}`
+	# Update battery volt data.
+	UPS_BATTERY_VOLT=`cut -c 3-6 < ${UPS_TTY_DATA}`
 }
 
 #/*---------------------------------------------------------------------------*/
 function read_ups_status {
-	# ttyACM response data background setup.
+	# ttyACM response data wait settings.
 	cat ${UPS_TTY_NODE} > ${UPS_TTY_DATA} &
 
-	# Get PID(cat command) for kill background process
+	# Get PID (cat command) to kill background process
+	PID=""
 	PID=$!
 
-	# UPS read status command send
-	sleep 0.5
+	# Send command to read charger status to UPS.
+	sleep 1
 	echo -ne "@C0#" > ${UPS_TTY_NODE}
-	sleep 0.5
+	sleep 1
 
-	# Kill Background process(cat cmd)
-	kill $PID
+	# Kill background process(cat cmd)
+	if [ -n "$PID" ]; then
+		kill $PID
+	fi
 
-	# Update UPS Status data
+	# Update charger status data.
 	UPS_STATUS_CHRG=`cut -c 6 < ${UPS_TTY_DATA}`
 	UPS_STATUS_FULL=`cut -c 4 < ${UPS_TTY_DATA}`
 }
@@ -97,17 +181,26 @@ function check_ups_status {
 		echo "ERROR: Battery Removed. force power off..."
 		echo "------------------------------------------------------------"
 		system_poweroff
+		return
 	fi
 
 	#/* UPS Status : Discharging... */
 	if [ ${UPS_STATUS_CHRG} -eq "1" -a ${UPS_STATUS_FULL} -eq "1" ]; then
 		echo "UPS Battery Status : Discharging..."
-		echo "UPS Battery Volt = ${UPS_STATUS_VOLT} mV"
+		echo "UPS Battery Volt = ${UPS_BATTERY_VOLT} mV"
+
 		#/* UPS Battery Status : Low Battery */
-		if [ ${UPS_STATUS_VOLT} -lt ${POWER_OFF_VOLT} ]; then
-			echo "------------------------------------------------------------"
-			echo "UPS Battery Volt ${UPS_STATUS_VOLT} mV is less then Power Off Volt ${POWER_OFF_VOLT} mV"
-			echo "------------------------------------------------------------"
+		if [ ${UPS_BATTERY_VOLT} -lt ${POWER_OFF_VOLT} ]; then
+			if [ ${POWER_OFF_VOLT} -eq ${BAT_LEVEL_FULL} ]; then
+				#/* Power off after Detecting UPS battery discharge. */
+				echo "------------------------------------------------------------"
+				echo "Detected UPS battery discharge."
+				echo "------------------------------------------------------------"
+			else
+				echo "------------------------------------------------------------"
+				echo "Power off when UPS_BATTERY_VOLT is lower than POWER_OFF_VOLT."
+				echo "------------------------------------------------------------"
+			fi
 			system_poweroff
 		fi
 	else
@@ -116,25 +209,33 @@ function check_ups_status {
 		else
 			echo "UPS Battery Status : Full Charged."
 		fi
-		echo "UPS Battery Volt = ${UPS_STATUS_VOLT} mV"
+
+		if [ ${POWER_OFF_VOLT} -eq ${BAT_LEVEL_FULL} ]; then
+			echo "UPS Power OFF : Detecting UPS battery discharge."
+		else
+			echo "UPS Power OFF : Battery Volt = ${POWER_OFF_VOLT} mV"
+		fi
 	fi
 }
 
 #/*---------------------------------------------------------------------------*/
 function system_poweroff {
-	# ttyACM response data background setup.
+	# ttyACM response data wait settings.
 	cat ${UPS_TTY_NODE} > ${UPS_TTY_DATA} &
 
-	# Get PID(cat command) for kill background process
+	# Get PID (cat command) to kill background process
+	PID=""
 	PID=$!
 
-	# UPS Poweroff command send
-	sleep 0.5
+	# Send command to ups off to UPS.
+	sleep 1
 	echo -ne "@P0#" > ${UPS_TTY_NODE}
-	sleep 0.5
+	sleep 1
 
-	# Kill Background process(cat cmd)
-	kill -9 $PID 2>&1
+	# Kill background process(cat cmd)
+	if [ -n "$PID" ]; then
+		kill -9 $PID
+	fi
 
 	echo "------------------------------------------------------------"
 	echo "run poweroff command..."
@@ -149,7 +250,7 @@ function system_poweroff {
 find_tty_node
 
 #/*---------------------------------------------------------------------------*/
-#/* tty noode가 없는 경우 처리 (script exit) */
+#/* Script exit handling when node not found. */
 #/*---------------------------------------------------------------------------*/
 if [ -z "${UPS_TTY_NODE}" ]; then
 	echo "------------------------------------------------------------"
@@ -163,7 +264,7 @@ else
 fi
 
 #/*---------------------------------------------------------------------------*/
-#/* 강제종료시 발생되는 cat command의 dead process kill. */
+#/* Kill previously running processes. */
 #/*---------------------------------------------------------------------------*/
 kill_dead_process
 
@@ -178,11 +279,15 @@ stty -F ${UPS_TTY_NODE} 9600 raw -echo
 while true
 do
 	echo "------------------------------------------------------------"
-	read_ups_volt
+	# Send command to read battery volt to UPS.
+	UPS_CMD_STR=${UPS_CMD_BATTERY_VOLT}
+	ups_cmd_send
+
+#	read_ups_volt
 	read_ups_status
 	date
 	check_ups_status
-	echo "Power OFF Volt = ${POWER_OFF_VOLT} mV"
+	echo "UPS Battery Volt = ${UPS_BATTERY_VOLT} mV"
 	echo "------------------------------------------------------------"
 done
